+++ HISTORY.rst
@@ -5,4 +5,4 @@ History
 
 0.2.0 (2017-08-17)
 ++++++++++++++++++
-* add search/autocomplete api using elasticsearch backend (5.5)
+* add search/autocomplete api using elasticsearch backend (2.4)